requirements.txt
streamlit
fastapi
uvicorn
huggingface_hub
pypdf
beautifulsoup4
python-multipart
python-dotenv

